FF-2617 - moved non analyzer rules out of the main folder
