+++ src/FunFair.CodeAnalysis/Helpers/LiteralString.cs
@@ -1,7 +1,7 @@
 using System;
 using Microsoft.CodeAnalysis;
 
-namespace FunFair.CodeAnalysis
+namespace FunFair.CodeAnalysis.Helpers
 {
     internal sealed class LiteralString : LocalizableString
     {
