+++ src/FunFair.CodeAnalysis/Helpers/RuleHelpers.cs
@@ -1,6 +1,6 @@
 using Microsoft.CodeAnalysis;
 
-namespace FunFair.CodeAnalysis
+namespace FunFair.CodeAnalysis.Helpers
 {
     internal static class RuleHelpers
     {
