+++ src/FunFair.CodeAnalysis/Helpers/Rules.cs
@@ -1,4 +1,4 @@
-namespace FunFair.CodeAnalysis
+namespace FunFair.CodeAnalysis.Helpers
 {
     internal static class Rules
     {
